Add files via upload
Great job at ODSC! I really enjoyed your talk on mapping data in R. I did have suggestions to add to your code. After your talk I mentioned the addResetButton(). This is a great option that resets the map with a click if you move away from the map. Also I changed the basemaps to a overlaygroups. The way you did it works if you prefer the radio buttons, but personally I like this as it allows more control, and you do not have to addCircles for each group.